// src/generated/models/User.ts
export type UserSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  name?: boolean
  email?: boolean
  emailVerified?: boolean
  image?: boolean
  address?: boolean
  phone?: boolean
  status?: boolean
  timezone?: boolean
  times?: boolean
  stripeCustomerId?: boolean
  createdAt?: boolean
  updatedAt?: boolean
  accounts?: boolean | Prisma.User$accountsArgs<ExtArgs>
  sessions?: boolean | Prisma.User$sessionsArgs<ExtArgs>
  Authenticator?: boolean | Prisma.User$AuthenticatorArgs<ExtArgs>
  services?: boolean | Prisma.User$servicesArgs<ExtArgs>
  reminders?: boolean | Prisma.User$remindersArgs<ExtArgs>
  appointments?: boolean | Prisma.User$appointmentsArgs<ExtArgs>
  subscription?: boolean | Prisma.User$subscriptionArgs<ExtArgs>
  _count?: boolean | Prisma.UserCountOutputTypeDefaultArgs<ExtArgs>
}, ExtArgs["result"]["user"]>
// ...
export type UserOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "name" | "email" | "emailVerified" | "image" | "address" | "phone" | "status" | "timezone" | "times" | "stripeCustomerId" | "createdAt" | "updatedAt", ExtArgs["result"]["user"]>
export type UserInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  accounts?: boolean | Prisma.User$accountsArgs<ExtArgs>
  sessions?: boolean | Prisma.User$sessionsArgs<ExtArgs>
  Authenticator?: boolean | Prisma.User$AuthenticatorArgs<ExtArgs>
  services?: boolean | Prisma.User$servicesArgs<ExtArgs>
  reminders?: boolean | Prisma.User$remindersArgs<ExtArgs>
  appointments?: boolean | Prisma.User$appointmentsArgs<ExtArgs>
  subscription?: boolean | Prisma.User$subscriptionArgs<ExtArgs>
  _count?: boolean | Prisma.UserCountOutputTypeDefaultArgs<ExtArgs>
}
export type UserIncludeCreateManyAndReturn<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {}
export type UserIncludeUpdateManyAndReturn<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {}

export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    accounts: Prisma.$AccountPayload<ExtArgs>[]
    sessions: Prisma.$SessionPayload<ExtArgs>[]
    Authenticator: Prisma.$AuthenticatorPayload<ExtArgs>[]
    services: Prisma.$ServicePayload<ExtArgs>[]
    reminders: Prisma.$ReminderPayload<ExtArgs>[]
    appointments: Prisma.$AppointmentPayload<ExtArgs>[]
    subscription: Prisma.$SubscriptionPayload<ExtArgs> | null
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string | null
    email: string
    emailVerified: Date | null
    image: string | null
    address: string | null
    phone: string | null
    status: boolean
    timezone: string | null
    times: string[]
    stripeCustomerId: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["user"]>
  composites: {}
}

export type UserGetPayload<S extends boolean | null | undefined | UserDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$UserPayload, S>
// ...
export type UserDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the User
   */
  select?: Prisma.UserSelect<ExtArgs> | null
  /**
   * Omit specific fields from the User
   */
  omit?: Prisma.UserOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.UserInclude<ExtArgs> | null
}